Consolidate (most) logging functionality in a single module
