+++ ratatelm/src/lib.rs
@@ -1,6 +1,7 @@
 //! `ratatelm` is an [elm](https://guide.elm-lang.org/architecture/)-like framework based on [ratatui](https://ratatui.rs/)
 
 mod tui;
+mod log;
 /// Contains the `Widget` trait and all customs widgets defined in ratatelm
 pub mod widgets;
 
@@ -11,37 +12,6 @@ use ratatui::prelude::Frame;
 
 use crossterm::event::{self, Event, KeyEvent};
 use std::time::Duration;
-
-// TODO these should be #cfg[(debug_assertions)]
-use std::sync::Mutex;
-lazy_static::lazy_static! {
-    static ref LOGS: Mutex<Vec<Event>> = Vec::new().into();
-
-    static ref LOGFILE_NAME: Mutex<String> = {
-        let mut sys = sysinfo::System::new();
-        sys.refresh_all();
-
-        format!(
-            "./logs/{}.log",
-            sys
-                .process(sysinfo::get_current_pid().expect("Should be able to current current pid"))
-                .expect("Should be able to get current process")
-                .start_time()
-        )
-    }.into();
-}
-
-/// # Panics
-/// Panics if unable to write to file or access system process information 
-pub fn dump_logs() { // TODO this shouldn't actually be public beyond the crate
-    std::fs::write(
-        LOGFILE_NAME
-            .lock()
-            .expect("The logfile name should have been initialized and not yet consumed")
-            .clone(),
-        format!("{:#?}", LOGS.lock().unwrap())
-    ).expect("Should be able to write to file");
-}
 
 /// The `App` trait is the entry point to a `ratatelm` application.
 ///
@@ -159,7 +129,7 @@ pub trait App <Message> {
             // Handle events and map to a Message
             if let Some(event) = Self::get_event()?{
                 #[cfg(debug_assertions)] {
-                    log(&event);
+                    log::log(&event);
                 }
 
                 self.handle_event(event)?;
@@ -169,20 +139,8 @@ pub trait App <Message> {
         tui::restore_terminal()?;
 
         #[cfg(debug_assertions)]
-        dump_logs();
+        log::dump_logs();
 
         Ok(())
     }
-}
-
-/// Records the event to a logfile for use in later debugging
-#[cfg(debug_assertions)]
-fn log(event: &Event) {
-    LOGS.lock().unwrap().push(event.clone());
-}
-
-/// Reads the next event back from the logfile
-#[cfg(debug_assertions)]
-#[must_use] fn delog() -> Event {
-    todo!()
 }
