+++ ratatelm/src/tui.rs
@@ -9,7 +9,7 @@ use ratatui::prelude::*;
 use std::{io::stdout, panic};
 use anyhow::Result;
 
-use crate::dump_logs;
+use crate::log::dump_logs;
 
 pub fn init_terminal() -> Result<Terminal<impl Backend>> {
     enable_raw_mode()?;
@@ -24,11 +24,13 @@ pub fn restore_terminal() -> Result<()> {
     Ok(())
 }
 
+
+
 pub fn install_panic_hook() {
     let original_hook = panic::take_hook();
     panic::set_hook(Box::new(move |panic_info| {
-        stdout().execute(LeaveAlternateScreen).unwrap();
-        disable_raw_mode().unwrap();
+        restore_terminal().unwrap();
+        // TODO dealing with logging really shouldn't be in the tui module
         #[cfg(debug_assertions)]
         dump_logs();
         original_hook(panic_info);
